requirements.txt
asgiref==3.11.0
attrs==25.4.0
dj-database-url==3.1.0
Django==6.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
drf-spectacular==0.26.5
gunicorn==21.2.0
inflection==0.5.1
jsonschema==4.17.3
packaging==26.0
pillow==10.2.0
psycopg2-binary==2.9.9
PyPDF2==3.0.1
pyrsistent==0.20.0
pytesseract==0.3.13
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.5
tzdata==2025.3
uritemplate==4.2.0
whitenoise==6.6.0
